requirements.txt
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==36.0.2
fastapi==0.88.0
fastapi-another-jwt-auth==0.1.6
h11==0.14.0
idna==3.6
jose==1.0.0
orjson==3.9.12
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.22.0
typing_extensions==4.9.0
uvicorn==0.20.0
